% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fat_stemness_riskscore.R
\name{fat_stemness_BCR_Score}
\alias{fat_stemness_BCR_Score}
\title{Calculate Samples BCR_Score with Fatty Acid Metabolism and Stemness gene list}
\usage{
fat_stemness_BCR_Score(testdata, scale_data = TRUE)
}
\arguments{
\item{testdata}{testdata is expression dataframe,
But the rownames are samples names,the colnames are gene names.}

\item{scale_data}{scale input data}
}
\value{
A BCR_Score of samples
}
\description{
Calculate Samples BCR_Score with Fatty Acid Metabolism and Stemness gene list
}
\section{Functions}{
\itemize{
\item \code{fat_stemness_BCR_Score()}: function is used to calculate sample BCR_Score
based on fatty acid metabolism and stemness gene list.

}}
\examples{
\donttest{
scores <- fat_stemness_BCR_Score(test_input_data)
scores
}

}
